+++ packages/integrations/vercel/package.json
@@ -23,7 +23,8 @@
     "./serverless/entrypoint": "./dist/serverless/entrypoint.js",
     "./static": "./dist/static/adapter.js",
     "./analytics": "./dist/analytics.js",
-    "./image-service": "./dist/image-service.js",
+    "./build-image-service": "./dist/image/build-service.js",
+    "./dev-image-service": "./dist/image/dev-service.js",
     "./package.json": "./package.json"
   },
   "typesVersions": {
