+++ packages/integrations/vercel/src/image/dev-service.ts
@@ -1,0 +1,34 @@
+import type { LocalImageService } from 'astro';
+import { sharedValidateOptions } from './shared';
+
+const service: LocalImageService = {
+	validateOptions: (options) => sharedValidateOptions(options, {}, 'development'),
+	getURL(options) {
+		const fileSrc = typeof options.src === 'string' ? options.src : options.src.src;
+
+		const searchParams = new URLSearchParams();
+		searchParams.append('url', fileSrc);
+
+		options.width && searchParams.append('w', options.width.toString());
+		options.quality && searchParams.append('q', options.quality.toString());
+
+		return '/_image' + searchParams;
+	},
+	parseURL(url) {
+		const params = url.searchParams;
+
+		if (!params.has('href')) {
+			return undefined;
+		}
+
+		const transform = {
+			src: params.get('href')!,
+			width: params.has('w') ? parseInt(params.get('w')!) : undefined,
+			quality: params.get('q'),
+		};
+
+		return transform;
+	},
+};
+
+export default service;
